refactor: 테스트케이스 네이밍 수정
