+++ app/src/androidTest/java/nextstep/shoppingcart/shoppinglist/shoppinglist/navigation/ShoppingListScreenNavigationTest.kt
@@ -41,11 +41,10 @@ class ShoppingListScreenNavigationTest {
     }
 
     @Test
-    fun 상품을_클릭하면_상품_상세보기_뷰로_이동한다() {
+    fun 상품을_클릭할_경우_상품상세보기_뷰로_이동한다1() {
         // given:
         // when:
-        composeTestRule.onAllNodesWithContentDescription("ShoppingItem").onFirst()
-            .performClick()
+        composeTestRule.onAllNodesWithContentDescription("ShoppingItem").onFirst().performClick()
 
         // then:
         val actual = navController.currentBackStackEntry?.destination?.route
@@ -55,7 +54,20 @@ class ShoppingListScreenNavigationTest {
     }
 
     @Test
-    fun 장바구니_아이콘을_클릭하면_장바구니_뷰로_이동한다() {
+    fun 상품을_클릭할_경우_상품상세보기_뷰로_이동한다2() {
+        // given:
+        // when:
+        composeTestRule.onAllNodesWithContentDescription("ShoppingItem")[2].performClick()
+
+        // then:
+        val actual = navController.currentBackStackEntry?.destination?.route
+        val expected = "ShoppingDetailScreen/{productId}"
+
+        assert(actual == expected)
+    }
+
+    @Test
+    fun 장바구니_아이콘을_클릭할_경우_장바구니_뷰로_이동한다() {
         // given:
         // when:
         composeTestRule.onNodeWithContentDescription("ShoppingCartIcon").performClick()
